Separate Closure from LLVM-related bits.
