+++ src/CMakeLists.txt
@@ -208,6 +208,7 @@ set(HEADER_FILES
   BoundsInference.h
   Buffer.h
   CSE.h
+  Closure.h
   CodeGen_ARM.h
   CodeGen_C.h
   CodeGen_GPU_Dev.h
@@ -349,7 +350,7 @@ add_library(Halide ${HALIDE_LIBRARY_TYPE}
   Bounds.cpp
   BoundsInference.cpp
   Buffer.cpp
-  CSE.cpp
+  Closure.cpp
   CodeGen_ARM.cpp
   CodeGen_C.cpp
   CodeGen_GPU_Dev.cpp
@@ -367,6 +368,7 @@ add_library(Halide ${HALIDE_LIBRARY_TYPE}
   CodeGen_Posix.cpp
   CodeGen_Renderscript_Dev.cpp
   CodeGen_X86.cpp
+  CSE.cpp
   Debug.cpp
   Debug.cpp
   DebugToFile.cpp
