+++ src/CodeGen_LLVM.cpp
@@ -2923,14 +2923,14 @@ void CodeGen_LLVM::visit(const For *op) {
 
         // Find every symbol that the body of this loop refers to
         // and dump it into a closure
-        Closure closure(op->body, op->name, buffer_t_type);
+        Closure closure(op->body, op->name);
 
         // Allocate a closure
-        StructType *closure_t = closure.build_type(context);
+        StructType *closure_t = build_closure_type(closure, buffer_t_type, context);
         Value *ptr = create_alloca_at_entry(closure_t, 1);
 
         // Fill in the closure
-        closure.pack_struct(closure_t, ptr, symbol_table, builder);
+        pack_closure(closure_t, ptr, closure, symbol_table, buffer_t_type, builder);
 
         // Make a new function that does one iteration of the body of the loop
         llvm::Type *voidPointerType = (llvm::Type *)(i8->getPointerTo());
@@ -2976,7 +2976,7 @@ void CodeGen_LLVM::visit(const For *op) {
         Value *closure_handle = builder->CreatePointerCast(iterator_to_pointer(iter),
                                                            closure_t->getPointerTo());
         // Load everything from the closure into the new scope
-        closure.unpack_struct(symbol_table, closure_t, closure_handle, builder);
+        unpack_closure(closure, symbol_table, closure_t, closure_handle, builder);
 
         // Generate the new function body
         codegen(op->body);
